Hotfix for async matching for isolation-group redirection (#5423)
What changed?

We observed some problems in production where a heavily backlogged backfill event got stuck and tasks were not dispatched. The issue was mitigated by rebooting matching, but more so by turning off zonal isolation for this one service.

Context
When a task is attempting to be dispatched through async dispatch, but it times out and there doesn't appear to be any pollers in the new zone, the feature will attempt to redirect the traffic to an isolation group where there are pollers - aka 'task redispatch'. This is intended to be a feature to avoid traffic from getting blackholed, but also to prevent it from blocking progression for other isolation groups in the tasklist, since it's a FIFO queue.

Suspected problem:
The suspicion is that the task processing is failing due to getting to a deadlock around task redispatch.

This pull-request attempts to fix the suspected problem of isolation-group task redirection causing a deadlock between the task pump and task dispatch. The basic idea being there's a few scenarios where it's conceivable that redispatching might block on putting tasks into that channel, and so it needs to be made nonblocking (all credit to Zijian/@Shaddoll for thinking of this / Steven for suggesting that also invalid channels should be checked)

Changes
- Refactors async dispatch code into two sub-functions to make it easier to read
- Adds some tests around blocking behaviour
- Adds default on select to prevent this from blocking (the actual bugfix)
- Adds check on redirection to ensure there's no invalid channel pushing, since that will also block, I'm told (til)

How did you test it?
Unit tests, basic load tests, integration tests