Improves history handler error metrics and logs (#5438)
Some bad data caused a tasklist to get stuck while trying to update a workflow which was in a bad state. This was more difficult to identify than it should be because the RunID was missing information from the error logs. This change includes this additional information and also protects some of the metrics emission from wrapped errors being misclassified.

This is not an attempt to do all error handling / wrapping correctly, it's just an incremental change, of which many more are needed.