Warn about bug introduced in 1.2.5 (#5449)
#5361 unfortunately broke timer-fixers, as the shared workflow requires a non-empty "enabled" list and none was provided in that PR.

#5433 restores that functionality.

---

Arguably `v1.2.5` could be retracted, as it introduces a potentially-problematic bug.

Since this fixer is disabled by default, we're going to skip doing that: we would have to publish `v1.2.6` to put the retraction into effect, and anyone updating that rapidly will not have much down-time.  Plus the fixer is not required for correct behavior.