Submodules step 1/2: convert archivers to a plugin architecture (#5597)
Breaking the gcloud plugin into a separate submodule (so the main module can remove these optional libraries) requires that the main module does not import it.

Since this was a hard-coded switch statement, that wasn't really possible.  So I've converted it to an at-`init`-time registration system, and I'll invert that import-flow in a later PR.  This one is intentionally trying to be a (relatively) straightforward refactor.

This is almost certainly better not using `init`, and the scheme/key separation seems strange... but that's a less-obviously-identical-behavior change, and we kinda need to split this apart ASAP.
Step 2 will introduce two submodules (gcloud plugin, and cmd/server which will import both the plugin and the "main" module), along with some makefile/automation changes to make things continue working like before.