Global ratelimiter, part 2: any-typed RPCs, mappers, and stub handler
# Any type:
This API is purely for inter-Cadence-service use, so the Thrift equivalent is in the github.com/uber/cadence-idl repository.

Design-wise: this has gone through several iterations, and nearly all top-level keys I can come up with are either totally insufficient or totally unnecessary for one ratelimit design or another, so I've leaned more and more towards "define nothing, allow passing anything".

Obviously this is not super ergonomic, but as the ratelimiter is intended to be pluggable by outside implementations it cannot *wholly* be defined in internal protocol definitions.
Implementers need to maintain in-process definitions on both ends of the protocol, and check type IDs before decoding.  Generally speaking this is probably best done with either a shared protobuf definition (anywhere) or something schema-free like JSON.

# Mappers
As the RPC specs and the internal type do essentially no interpreting of their contents, the mappers are both trivial and complete.

# Handler
Fully wiring this up will come later, and at a very high level will require:
- a `common/quota` wrapper to collect usage metrics in the "limiting" hosts (frontend)
- periodically, send bulk `RatelimitUpdateRequest`s to "aggregating" hosts (history)
- use the responses to update quotas as needed
- dynamicconfig to enable/disable/shadow/roll-back all this

All of which is essentially out of scope for this PR.