persistence/client/factory.go test coverage (#5859)
Our ability to cover this file is rather harshly limited due to it directly calling funcs in other packages that try to make network connections... but thankfully the DataStore-based stuff can be intercepted.  So I've at least tested those branches.

More generally this package is yet another strong argument in favor of leaning more on go.uber.org/fx for dependency injection.  Very little of this has to be (or should be) hard-coded in this file.  It feels like Java's "Service Locator" pattern, but even worse because the language doesn't allow test controls of globals like Java does... so I doubt it's a coincidence that there's a `bean.go` file in here.