We now wait 10 seconds before we start returning shard closed errors, also stop retrying on shard closed errors  (#5938)
What changed?
- Introduced a shardRecentlyClosed error to signal from the shard that it was recently closed. This error will not cause the task handler to emit error logs and metrics.
- Add a test that hits this guard in all the places it exists
- Removed redundant tests that checked the same as the new guard test now checks
- Stop retrying on shard closed errors

Why?
Shard closing is not an unexpected state, so we should not emit error logs and metrics for this.

If we stay in a state where a closed shard keeps getting requests, then we should start emitting error logs and metrics, so we wait 10 seconds, and if we still see requests then we start emitting the metrics.

All the guard testing and deleting is necessary to make the new line coverage check happy.

How did you test it?
Tested with unit tests and by deploying to staging. The deployment shows we can now do restarts without seeing these errors.

Potential risks
This does change some relatively core task processing logic, however the main change is on how the error states are communicated. The main flow is not touched.

Release notes

Documentation Changes