Bugfix: we address hosts using string(rune(shardID)), not by itoa(shardD) (#5952)
* Bugfix: we address hosts using string(rune(shardID)), not by
itoa(shardD)

I wish it be the other way, but we can't change how it works now since
in all the other places we already use rune(shardID).
This was not a critical issue - we just returned ShardOwnershipLostError
with bad target host in very specific cases. For instance, at the
persistence engine error - when we got an error from transaction.

Extracting the weird string(rune(shardID)) logic to a separate function
so to prevent the same issue in future.