Enforcing go vet -copylocks and fixing current violations (#5967)
Thankfully this all appears to be test-related and not any real risks,
but I realized while writing some other code that `go test` doesn't check this.

I'm really not sure what scenarios it might be wrong in, but it catches loads
of incorrect concurrent code in my experience.  So now it's required.