We now only log "Failed to sig ..." on unexpected errors
We have a log of "Failed to signal extermal workflow execution" error
logs in server with the error being "workflow execution already
completed".

This is an expected state so we should not log errors for this.
Additionally we add transient errors and not exists errors to the things we
should not log errors for, as they are expected in the normal operation
of the system.