Fix rebalancing failover tooling
**What changed?**
Excluded `domains` that have `preferredClusters` that are not present in its `cluster` list in `ReplicationConfiguration`

**Why?**
`Domains` with `preferredClusters` not in their `cluster` list were causing the `workflow` to panic when trying to acquire the `client` for the `preferredCluster` from the `remoteFrontendClients`

**How did you test it?**
unit tests and local replication tests

**Potential risks**
The `workflow` was generally not working before because having a `preferredCluster` not in the `domain` `cluster` list is a normal scenario. The risk introduced is the `workflow` actually executing and doing something unexpected. For now, we just rebalance `domains` that have `preferredCluster` set and that can be rebalanced back to their `preferredCluster`

**Release notes**

**Documentation Changes**