Switch quotas.Collection to clock.Ratelimiter, minor fixes along the way
Partly this is worth doing at some point regardless, and partly this is a
bit of a blocker for the global ratelimiter, as it needs to expose some
Reservation-using APIs to fit into existing ratelimiter code.

This has a larger blast-radius than I'd like, as it's not just frontend,
but it should be safe - the clock.Ratelimiter is pretty thoroughly tested.

If we are feeling paranoid though, I could split this collection API into
two, so frontend switches to clock.Ratelimiter and matching can remain
untouched.  A good chunk of the quotas folder is tightly intertwined though,
and some seems very unnecessary, so I'm a bit reluctant to do that unless
someone's worried.