bugfix: bad dynamicconfig filter/string mapping (#6151)
This logic desperately needs to be refactored, it's incredibly error-prone :\
We should probably just use enumer's codegen tbh.  Or something similar.

Previously `String()` missed both `workflowType` and `ratelimitKey`.
The `String()` impl is now rewritten so it won't be missed with future additions, and there's a test to check it too.