Ratelimiter usage-counting bugfix: rejected reservations were not counted (#6158)
After noticing that we had rejected requests but no rejected global-limit numbers, I noticed this flaw.
Basically it's just that not-allowed requests were not being counted in the fallback-limiter, so it only affected the global system.

Since "counted" and "fallback" have separate implementations, and reserve is easier to mess up (as demonstrated) but rather important because it's heavily used in the frontend per-domain requests: I've added a test to check all three things to make sure they track calls like they should.

As part of this, to share some types a bit more and to just normalize the tests a bit, I dropped the external-test-package part of the fallback-limiter test.  AFAICT it was working fine, but it's definitely a bit of an abnormality in this codebase, and I'm not confident that it works correctly with coverage systems (I just haven't checked).  Might as well simplify.

---

In terms of concrete changes, there are really only two here:
- fallback-limiter's Reserve should have called Used(false) and counted the rejection.
- "not idle" is now done at `Used(..)` time, not `Reserve()`

The rest is tests and minor refactoring to simplify it as much as possible.