Add additional options to matching simulation cases
Add RecordDecisionTaskStartedTime to control the delay incurred by matching calling to history.

Add TaskProcessTime to control the delay in-between polls. This is meant to represent time the client spends executing the task.

Add TasksBurst to allow setting the burst value on the rate limiter, defaulting to a value of 1. Currently it uses the same value as the RPS which results in spikey behavior that causes sync matching to significantly underperform.

Allow ForwarderMaxOutstandingPolls and ForwarderMaxOutstandingTasks to be zero. It's interesting to see a control without any forwarding.

Additionally add a few test cases to showcase different behaviors of the current matching stack. This isn't a complete set yet but each of these brings some interesting behavior.