Change dockerhub user from ubercadence to jht305
The ubercadence dockerhub user has been changed from a user to an
organization, this means we can no longer log in as ubercadence.

Sadly dockerhub does not support access tokens for organisations, see
https://github.com/docker/roadmap/issues/461 it is however on the
roadmap https://github.com/orgs/docker/projects/51

Until this is supported the workaround is to use a user login that has
access to the org. So as I am such a user we will use mine for the time
being.