Add metrics for tasks already started in matching
What changed?
Added metrics to PollForDecisionTask and PollForActivityTask to track when the task being polled has already started and therefore discarded.

Why?
This metric will give visibility to the accumulation of stale tasks in matching and will serve as a baseline for future work on that part.

How did you test it?

Potential risks

Release notes

Documentation Changes