Expose the current Limit() on existing ratelimiters (#6235)
A rather minor change, but it will be useful for a later PR:
I'm building some logic into the global ratelimiter that needs to know what the "local" limit would be, because that value is assumed to be "safe" in the absence of other info.
The easiest and IMO cleanest way to implement this is to allow interrogating the existing limiters (the global-fallback one specifically), rather than re-implementing the "get the limit and divide by the ring size" logic separately, because that logic may drift.

More generally: I think we'll eventually either get rid of `quotas.Limiter`, or turn it into a complete read-only subset of `clock.Ratelimiter` (because *very* few things need to write to limits).  This is an incremental step towards that.

I also attempted to change all the other things touching this to `rate.Limit`, but... that turned out to be rather large.  I think it's worth doing, primitive types are unnecessarily vague in nearly all cases, but not right now.